+++ state-chain/cf-integration-tests/src/account.rs
@@ -84,7 +84,7 @@ fn account_deletion_removes_relevant_storage_items() {
 		assert!(AccountPeerMapping::<Runtime>::get(&backup_node).is_none());
 		assert!(!MappedPeers::<Runtime>::contains_key(peer_id));
 		let vanity_names = VanityNames::<Runtime>::get();
-		assert!(vanity_names.get(&backup_node).is_none());
+		assert!(!vanity_names.contains_key(&backup_node));
 		assert_eq!(pallet_cf_account_roles::AccountRoles::<Runtime>::get(&backup_node), None);
 		assert_eq!(Reputations::<Runtime>::get(backup_node).online_blocks, 0);
 	});