+++ state-chain/custom-rpc/src/lib.rs
@@ -8,20 +8,21 @@ use cf_chains::{
 	address::{ForeignChainAddressHumanreadable, ToHumanreadableAddress},
 	dot::PolkadotAccountId,
 	eth::Address as EthereumAddress,
+	sol::SolAddress,
 	Chain,
 };
 use cf_primitives::{
 	chains::assets::any::{self, AssetMap},
 	AccountRole, Asset, AssetAmount, BlockNumber, BroadcastId, EpochIndex, ForeignChain,
-	NetworkEnvironment, SemVer, SwapId,
+	NetworkEnvironment, SemVer, SwapId, SwapRequestId,
 };
 use cf_utilities::rpc::NumberOrHex;
 use core::ops::Range;
 use jsonrpsee::{
-	core::{error::SubscriptionClosed, RpcResult},
+	core::RpcResult,
 	proc_macros::rpc,
-	types::error::{CallError, SubscriptionEmptyError},
-	SubscriptionSink,
+	types::error::{ErrorObject, ErrorObjectOwned},
+	PendingSubscriptionSink,
 };
 use order_fills::OrderFills;
 use pallet_cf_governance::GovCallHash;
@@ -40,9 +41,9 @@ use state_chain_runtime::{
 	chainflip::{BlockUpdate, Offence},
 	constants::common::TX_FEE_MULTIPLIER,
 	monitoring_apis::{
-		AuthoritiesInfo, BtcUtxos, EpochState, ExternalChainsBlockHeight, FeeImbalance, FlipSupply,
-		LastRuntimeUpgradeInfo, MonitoringData, OpenDepositChannels, PendingBroadcasts,
-		PendingTssCeremonies, RedemptionsInfo,
+		ActivateKeysBroadcastIds, AuthoritiesInfo, BtcUtxos, EpochState, ExternalChainsBlockHeight,
+		FeeImbalance, FlipSupply, LastRuntimeUpgradeInfo, MonitoringData, OpenDepositChannels,
+		PendingBroadcasts, PendingTssCeremonies, RedemptionsInfo, SolanaNonces,
 	},
 	runtime_apis::{
 		BoostPoolDepth, BoostPoolDetails, BrokerInfo, CustomRuntimeApi, DispatchErrorWithMessage,
@@ -61,7 +62,7 @@ use std::{
 pub mod monitoring;
 pub mod order_fills;
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct RpcEpochState {
 	pub blocks_per_epoch: u32,
 	pub current_epoch_started_at: u32,
@@ -80,7 +81,7 @@ impl From<EpochState> for RpcEpochState {
 		}
 	}
 }
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct RpcRedemptionsInfo {
 	pub total_balance: NumberOrHex,
 	pub count: u32,
@@ -93,7 +94,7 @@ impl From<RedemptionsInfo> for RpcRedemptionsInfo {
 
 pub type RpcFeeImbalance = FeeImbalance<NumberOrHex>;
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct RpcFlipSupply {
 	pub total_supply: NumberOrHex,
 	pub offchain_supply: NumberOrHex,
@@ -107,7 +108,7 @@ impl From<FlipSupply> for RpcFlipSupply {
 	}
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct RpcMonitoringData {
 	pub external_chains_height: ExternalChainsBlockHeight,
 	pub btc_utxos: BtcUtxos,
@@ -123,6 +124,10 @@ pub struct RpcMonitoringData {
 	pub pending_swaps: u32,
 	pub dot_aggkey: PolkadotAccountId,
 	pub flip_supply: RpcFlipSupply,
+	pub sol_aggkey: SolAddress,
+	pub sol_onchain_key: SolAddress,
+	pub sol_nonces: SolanaNonces,
+	pub activating_key_broadcast_ids: ActivateKeysBroadcastIds,
 }
 impl From<MonitoringData> for RpcMonitoringData {
 	fn from(monitoring_data: MonitoringData) -> Self {
@@ -141,6 +146,10 @@ impl From<MonitoringData> for RpcMonitoringData {
 			pending_swaps: monitoring_data.pending_swaps,
 			dot_aggkey: monitoring_data.dot_aggkey,
 			flip_supply: monitoring_data.flip_supply.into(),
+			sol_aggkey: monitoring_data.sol_aggkey,
+			sol_onchain_key: monitoring_data.sol_onchain_key,
+			sol_nonces: monitoring_data.sol_nonces,
+			activating_key_broadcast_ids: monitoring_data.activating_key_broadcast_ids,
 		}
 	}
 }
@@ -148,7 +157,7 @@ impl From<MonitoringData> for RpcMonitoringData {
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct ScheduledSwap {
 	pub swap_id: SwapId,
-	pub swap_request_id: SwapId,
+	pub swap_request_id: SwapRequestId,
 	pub base_asset: Asset,
 	pub quote_asset: Asset,
 	pub side: Side,
@@ -194,7 +203,7 @@ impl ScheduledSwap {
 	}
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct RpcLiquidityProviderBoostPoolInfo {
 	pub fee_tier: u16,
 	pub total_balance: U256,
@@ -225,7 +234,7 @@ impl From<&LiquidityProviderBoostPoolInfo> for RpcLiquidityProviderBoostPoolInfo
 }
 
 #[allow(clippy::large_enum_variant)]
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 #[serde(tag = "role", rename_all = "snake_case")]
 pub enum RpcAccountInfo {
 	Unregistered {
@@ -323,7 +332,7 @@ impl RpcAccountInfo {
 	}
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct RpcAccountInfoV2 {
 	pub balance: NumberOrHex,
 	pub bond: NumberOrHex,
@@ -340,7 +349,7 @@ pub struct RpcAccountInfoV2 {
 	pub restricted_balances: BTreeMap<EthereumAddress, u128>,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct RpcPenalty {
 	reputation_points: i32,
 	suspension_duration_blocks: u32,
@@ -348,7 +357,7 @@ pub struct RpcPenalty {
 
 type RpcSuspensions = Vec<(Offence, Vec<(u32, state_chain_runtime::AccountId)>)>;
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct RpcAuctionState {
 	blocks_per_epoch: u32,
 	current_epoch_started_at: u32,
@@ -358,7 +367,7 @@ pub struct RpcAuctionState {
 	min_active_bid: Option<NumberOrHex>,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct RpcSwapOutputV1 {
 	// Intermediary amount, if there's any
 	pub intermediary: Option<NumberOrHex>,
@@ -375,14 +384,14 @@ impl From<RpcSwapOutputV2> for RpcSwapOutputV1 {
 	}
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct RpcFee {
 	#[serde(flatten)]
 	pub asset: Asset,
 	pub amount: Amount,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct RpcSwapOutputV2 {
 	// Intermediary amount, if there's any
 	pub intermediary: Option<U256>,
@@ -411,12 +420,12 @@ impl From<PoolInfo> for RpcPoolInfo {
 	}
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct PoolsEnvironment {
 	pub fees: any::AssetMap<Option<RpcPoolInfo>>,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct IngressEgressEnvironment {
 	pub minimum_deposit_amounts: any::AssetMap<NumberOrHex>,
 	pub ingress_fees: any::AssetMap<Option<NumberOrHex>>,
@@ -426,22 +435,23 @@ pub struct IngressEgressEnvironment {
 	pub channel_opening_fees: HashMap<ForeignChain, NumberOrHex>,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct FundingEnvironment {
 	pub redemption_tax: NumberOrHex,
 	pub minimum_funding_amount: NumberOrHex,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct SwappingEnvironment {
 	maximum_swap_amounts: any::AssetMap<Option<NumberOrHex>>,
 	network_fee_hundredth_pips: Permill,
 	swap_retry_delay_blocks: u32,
 	max_swap_retry_duration_blocks: u32,
 	max_swap_request_duration_blocks: u32,
+	minimum_chunk_size: any::AssetMap<NumberOrHex>,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct RpcEnvironment {
 	ingress_egress: IngressEgressEnvironment,
 	swapping: SwappingEnvironment,
@@ -479,25 +489,25 @@ mod boost_pool_rpc {
 
 	use super::*;
 
-	#[derive(Serialize, Deserialize)]
+	#[derive(Serialize, Deserialize, Clone)]
 	struct AccountAndAmount {
 		account_id: AccountId32,
 		amount: U256,
 	}
 
-	#[derive(Serialize, Deserialize)]
+	#[derive(Serialize, Deserialize, Clone)]
 	struct PendingBoost {
 		deposit_id: PrewitnessedDepositId,
 		owed_amounts: Vec<AccountAndAmount>,
 	}
 
-	#[derive(Serialize, Deserialize)]
+	#[derive(Serialize, Deserialize, Clone)]
 	struct PendingWithdrawal {
 		account_id: AccountId32,
 		pending_deposits: BTreeSet<PrewitnessedDepositId>,
 	}
 
-	#[derive(Serialize, Deserialize)]
+	#[derive(Serialize, Deserialize, Clone)]
 	pub struct BoostPoolDetailsRpc {
 		fee_tier: u16,
 		#[serde(flatten)]
@@ -546,13 +556,13 @@ mod boost_pool_rpc {
 		}
 	}
 
-	#[derive(Serialize, Deserialize)]
+	#[derive(Serialize, Deserialize, Clone)]
 	struct PendingFees {
 		deposit_id: PrewitnessedDepositId,
 		fees: Vec<AccountAndAmount>,
 	}
 
-	#[derive(Serialize, Deserialize)]
+	#[derive(Serialize, Deserialize, Clone)]
 	pub struct BoostPoolFeesRpc {
 		fee_tier: u16,
 		#[serde(flatten)]
@@ -987,20 +997,26 @@ where
 	}
 }
 
-fn to_rpc_error<E: std::error::Error + Send + Sync + 'static>(e: E) -> jsonrpsee::core::Error {
-	CallError::from_std_error(e).into()
+pub fn str_to_rpc_error(e: &str) -> ErrorObjectOwned {
+	ErrorObject::owned(jsonrpsee::types::error::CALL_EXECUTION_FAILED_CODE, e, Option::<()>::None)
 }
 
-fn map_dispatch_error(e: DispatchErrorWithMessage) -> jsonrpsee::core::Error {
-	jsonrpsee::core::Error::from(match e {
-		DispatchErrorWithMessage::Module(message) => match std::str::from_utf8(&message) {
-			Ok(message) => anyhow::anyhow!("DispatchError: {message}"),
-			Err(error) =>
-				anyhow::anyhow!("DispatchError: Unable to deserialize error message: '{error}'"),
-		},
-		DispatchErrorWithMessage::Other(e) =>
-			anyhow::anyhow!("DispatchError: {}", <&'static str>::from(e)),
-	})
+pub fn to_rpc_error<E: core::fmt::Display + Send + Sync + 'static>(e: E) -> ErrorObjectOwned {
+	str_to_rpc_error(&format!("{}", e)[..])
+}
+
+fn map_dispatch_error(e: DispatchErrorWithMessage) -> ErrorObjectOwned {
+	str_to_rpc_error(
+		&(match e {
+			DispatchErrorWithMessage::Module(message) => match std::str::from_utf8(&message) {
+				Ok(message) => format!("DispatchError: {message}"),
+				Err(error) =>
+					format!("DispatchError: Unable to deserialize error message: '{error}'"),
+			},
+			DispatchErrorWithMessage::Other(e) =>
+				format!("DispatchError: {}", <&'static str>::from(e)),
+		})[..],
+	)
 }
 
 impl<C, B> CustomApiServer for CustomRpc<C, B>
@@ -1356,7 +1372,7 @@ where
 							Ok(amount)
 						}
 					})
-					.map_err(|str| anyhow::anyhow!(str))?,
+					.map_err(str_to_rpc_error)?,
 				additional_orders.map(|additional_orders| {
 					additional_orders
 						.into_iter()
@@ -1596,6 +1612,12 @@ where
 				.map_err(to_rpc_error)?,
 			max_swap_retry_duration_blocks: swap_limits.max_swap_retry_duration_blocks,
 			max_swap_request_duration_blocks: swap_limits.max_swap_request_duration_blocks,
+			minimum_chunk_size: any::AssetMap::try_from_fn(|asset| {
+				runtime_api
+					.cf_minimum_chunk_size(hash, asset)
+					.map_err(to_rpc_error)
+					.map(Into::into)
+			})?,
 		})
 	}
 
@@ -1661,28 +1683,28 @@ where
 
 	fn cf_subscribe_pool_price(
 		&self,
-		sink: SubscriptionSink,
+		pending_sink: PendingSubscriptionSink,
 		from_asset: Asset,
 		to_asset: Asset,
-	) -> Result<(), SubscriptionEmptyError> {
+	) {
 		self.new_subscription(
 			true,  /* only_on_changes */
 			false, /* end_on_error */
-			sink,
+			pending_sink,
 			move |client, hash| client.runtime_api().cf_pool_price(hash, from_asset, to_asset),
 		)
 	}
 
 	fn cf_subscribe_pool_price_v2(
 		&self,
-		sink: SubscriptionSink,
+		pending_sink: PendingSubscriptionSink,
 		base_asset: Asset,
 		quote_asset: Asset,
-	) -> Result<(), SubscriptionEmptyError> {
+	) {
 		self.new_subscription(
 			false, /* only_on_changes */
 			true,  /* end_on_error */
-			sink,
+			pending_sink,
 			move |client, hash| {
 				client
 					.runtime_api()
@@ -1696,23 +1718,23 @@ where
 
 	fn cf_subscribe_scheduled_swaps(
 		&self,
-		sink: SubscriptionSink,
+		pending_sink: PendingSubscriptionSink,
 		base_asset: Asset,
 		quote_asset: Asset,
-	) -> Result<(), SubscriptionEmptyError> {
+	) {
 		// Check that the requested pool exists:
 		let Ok(Ok(_)) = self.client.runtime_api().cf_pool_info(
 			self.client.info().best_hash,
 			base_asset,
 			quote_asset,
 		) else {
-			return Err(SubscriptionEmptyError);
+			return;
 		};
 
 		self.new_subscription(
 			false, /* only_on_changes */
 			true,  /* end_on_error */
-			sink,
+			pending_sink,
 			move |client, hash| {
 				Ok::<_, ApiError>(SwapResponse {
 					swaps: client
@@ -1723,7 +1745,7 @@ where
 						.collect(),
 				})
 			},
-		)
+		);
 	}
 
 	fn cf_scheduled_swaps(
@@ -1751,17 +1773,17 @@ where
 
 	fn cf_subscribe_prewitness_swaps(
 		&self,
-		sink: SubscriptionSink,
+		pending_sink: PendingSubscriptionSink,
 		base_asset: Asset,
 		quote_asset: Asset,
 		side: Side,
-	) -> Result<(), SubscriptionEmptyError> {
+	) {
 		self.new_subscription(
 			false, /* only_on_changes */
 			true,  /* end_on_error */
-			sink,
+			pending_sink,
 			move |client, hash| {
-				Ok::<_, jsonrpsee::core::Error>(RpcPrewitnessedSwap {
+				Ok::<_, ErrorObjectOwned>(RpcPrewitnessedSwap {
 					base_asset,
 					quote_asset,
 					side,
@@ -1799,10 +1821,7 @@ where
 		})
 	}
 
-	fn cf_subscribe_lp_order_fills(
-		&self,
-		sink: SubscriptionSink,
-	) -> Result<(), SubscriptionEmptyError> {
+	fn cf_subscribe_lp_order_fills(&self, sink: PendingSubscriptionSink) {
 		self.new_subscription_with_state(
 			false, /* only_on_changes */
 			true,  /* end_on_error */
@@ -1816,18 +1835,16 @@ where
 						let pools_events =
 							client.runtime_api().cf_lp_events(hash).map_err(to_rpc_error)?;
 
-						Ok::<_, jsonrpsee::core::Error>(
-							order_fills::order_fills_from_block_updates(
-								prev_pools,
-								&pools,
-								pools_events,
-							),
-						)
+						RpcResult::Ok(order_fills::order_fills_from_block_updates(
+							prev_pools,
+							&pools,
+							pools_events,
+						))
 					})
 					.transpose()?
 					.unwrap_or_default();
 
-				Ok::<_, jsonrpsee::core::Error>((fills, pools))
+				RpcResult::Ok((fills, pools))
 			},
 		)
 	}
@@ -1977,9 +1994,9 @@ where
 		&self,
 		only_on_changes: bool,
 		end_on_error: bool,
-		sink: SubscriptionSink,
+		sink: PendingSubscriptionSink,
 		f: F,
-	) -> Result<(), SubscriptionEmptyError> {
+	) {
 		self.new_subscription_with_state(
 			only_on_changes,
 			end_on_error,
@@ -2002,20 +2019,24 @@ where
 		&self,
 		only_on_changes: bool,
 		end_on_error: bool,
-		mut sink: SubscriptionSink,
+		pending_sink: PendingSubscriptionSink,
 		f: F,
-	) -> Result<(), SubscriptionEmptyError> {
-		use futures::{future::FutureExt, stream::StreamExt};
+	) {
+		use futures::{stream::StreamExt, FutureExt};
 
 		let info = self.client.info();
 
 		let (initial_item, initial_state) = match f(&self.client, info.best_hash, None) {
 			Ok(initial) => initial,
 			Err(e) => {
-				let _ = sink.reject(jsonrpsee::core::Error::from(
-					sc_rpc_api::state::error::Error::Client(Box::new(e)),
-				));
-				return Ok(())
+				self.executor.spawn(
+					"cf-rpc-update-subscription",
+					Some("rpc"),
+					pending_sink
+						.reject(sc_rpc_api::state::error::Error::Client(Box::new(to_rpc_error(e))))
+						.boxed(),
+				);
+				return;
 			},
 		};
 
@@ -2047,7 +2068,7 @@ where
 									data: new_item,
 								}))
 							},
-							Err(error) if end_on_error => Some(Err(error)),
+							Err(error) if end_on_error => Some(Err(to_rpc_error(error))),
 							_ => None,
 						})
 					}
@@ -2057,16 +2078,8 @@ where
 		self.executor.spawn(
 			"cf-rpc-update-subscription",
 			Some("rpc"),
-			async move {
-				if let SubscriptionClosed::Failed(err) = sink.pipe_from_try_stream(stream).await {
-					log::error!("Subscription closed due to error: {err:?}");
-					sink.close(err);
-				}
-			}
-			.boxed(),
+			sc_rpc::utils::pipe_from_stream(pending_sink, stream).boxed(),
 		);
-
-		Ok(())
 	}
 }
 
@@ -2246,6 +2259,18 @@ mod test {
 				swap_retry_delay_blocks: 5,
 				max_swap_retry_duration_blocks: 600,
 				max_swap_request_duration_blocks: 14400,
+				minimum_chunk_size: any::AssetMap {
+					eth: eth::AssetMap {
+						eth: 123_u32.into(),
+						flip: 0u32.into(),
+						usdc: 456_u32.into(),
+						usdt: 0u32.into(),
+					},
+					btc: btc::AssetMap { btc: 789_u32.into() },
+					dot: dot::AssetMap { dot: 0u32.into() },
+					arb: arb::AssetMap { eth: 0u32.into(), usdc: 101112_u32.into() },
+					sol: sol::AssetMap { sol: 0u32.into(), usdc: 0u32.into() },
+				},
 			},
 			ingress_egress: IngressEgressEnvironment {
 				minimum_deposit_amounts: any::AssetMap {