+++ state-chain/pallets/cf-elections/src/electoral_systems/composite.rs
@@ -31,6 +31,7 @@ mod tags {
 	pub struct C;
 	pub struct D;
 	pub struct EE;
+	pub struct FF;
 }
 
 macro_rules! generate_electoral_system_tuple_impls {
@@ -39,7 +40,7 @@ macro_rules! generate_electoral_system_tuple_impls {
         // We use the type names as variable names.
         #[allow(non_snake_case)]
         // In the 1/identity case, no invalid combinations are possible, so error cases are unreachable.
-        #[allow(unreachable_patterns)]
+
         // Macro expands tuples, but only uses 1 element in some cases.
         #[allow(unused_variables)]
         pub mod $module {
@@ -595,4 +596,4 @@ macro_rules! generate_electoral_system_tuple_impls {
     };
 }
 
-generate_electoral_system_tuple_impls!(tuple_5_impls: ((A, A0), (B, B0), (C, C0), (D, D0), (EE, E0)));
+generate_electoral_system_tuple_impls!(tuple_6_impls: ((A, A0), (B, B0), (C, C0), (D, D0), (EE, E0), (FF, F0)));