+++ state-chain/pallets/cf-elections/src/electoral_systems/tests.rs
@@ -4,6 +4,7 @@ pub(crate) use crate::register_checks;
 
 pub mod change;
 pub mod egress_success;
+pub mod liveness;
 pub mod monotonic_median;
 pub mod unsafe_median;
 pub mod utils;