+++ state-chain/pallets/cf-ingress-egress/src/mock_eth.rs
@@ -248,11 +248,16 @@ impl<Ctx: Clone> RequestAddress for TestExternalities<Test, Ctx> {
 				.cloned()
 				.map(|request| match request {
 					DepositRequest::Liquidity { lp_account, asset } =>
-						IngressEgress::request_liquidity_deposit_address(lp_account, asset, 0, None)
-							.map(|(id, addr, ..)| {
-								(request, id, TestChainAccount::try_from(addr).unwrap())
-							})
-							.unwrap(),
+						IngressEgress::request_liquidity_deposit_address(
+							lp_account,
+							asset,
+							0,
+							ForeignChainAddress::Eth(Default::default()),
+						)
+						.map(|(id, addr, ..)| {
+							(request, id, TestChainAccount::try_from(addr).unwrap())
+						})
+						.unwrap(),
 					DepositRequest::SimpleSwap {
 						source_asset,
 						destination_asset,