+++ state-chain/pallets/cf-threshold-signature/src/tests.rs
@@ -5,13 +5,13 @@ use crate::{
 	mock::*, AttemptCount, AuthorityCount, CeremonyContext, CeremonyId, CurrentEpochIndex, Error,
 	Event as PalletEvent, KeyHandoverResolutionPendingSince, KeyRotationStatus,
 	KeygenFailureVoters, KeygenOutcomeFor, KeygenResolutionPendingSince, KeygenResponseTimeout,
-	KeygenSuccessVoters, PalletOffence, PendingKeyRotation, RequestContext, RequestId,
-	ThresholdSignatureResponseTimeout,
+	KeygenSlashAmount, KeygenSuccessVoters, PalletConfigUpdate, PalletOffence, PendingKeyRotation,
+	RequestContext, RequestId, ThresholdSignatureResponseTimeout,
 };
 
 use cf_chains::mocks::{MockAggKey, MockEthereumChainCrypto};
-use cf_primitives::GENESIS_EPOCH;
-use cf_test_utilities::{last_event, maybe_last_event};
+use cf_primitives::{FlipBalance, GENESIS_EPOCH};
+use cf_test_utilities::last_event;
 use cf_traits::{
 	mocks::{
 		cfe_interface_mock::{MockCfeEvent, MockCfeInterface},
@@ -27,7 +27,7 @@ use frame_support::{
 	assert_err, assert_noop, assert_ok,
 	instances::Instance1,
 	pallet_prelude::DispatchResult,
-	traits::{Hooks, OnInitialize},
+	traits::{Hooks, OnInitialize, OriginTrait},
 };
 use frame_system::pallet_prelude::BlockNumberFor;
 use sp_runtime::traits::BlockNumberProvider;
@@ -1772,26 +1772,6 @@ mod key_rotation {
 }
 
 #[test]
-fn set_keygen_response_timeout_works() {
-	new_test_ext_no_key().execute_with(|| {
-		let init_timeout = KeygenResponseTimeout::<Test, _>::get();
-
-		EvmThresholdSigner::set_keygen_response_timeout(RuntimeOrigin::root(), init_timeout)
-			.unwrap();
-
-		assert!(maybe_last_event::<Test>().is_none());
-
-		let new_timeout = init_timeout + 1;
-
-		EvmThresholdSigner::set_keygen_response_timeout(RuntimeOrigin::root(), new_timeout)
-			.unwrap();
-
-		assert_last_events!(crate::Event::KeygenResponseTimeoutUpdated { .. });
-		assert_eq!(KeygenResponseTimeout::<Test, _>::get(), new_timeout)
-	});
-}
-
-#[test]
 fn dont_slash_in_safe_mode() {
 	new_test_ext().execute_with(|| {
 		MockRuntimeSafeMode::set_safe_mode(MockRuntimeSafeMode {
@@ -1964,19 +1944,59 @@ fn can_recover_from_abort_key_rotation_after_key_handover_failed() {
 }
 
 #[test]
-fn ensure_governance_origin_checks() {
+fn can_update_all_config_items() {
 	new_test_ext().execute_with(|| {
-		assert_noop!(
-			EvmThresholdSigner::set_keygen_response_timeout(RuntimeOrigin::signed(ALICE), 1),
-			sp_runtime::traits::BadOrigin,
+		const NEW_THRESHOLD_SIGNATURE_RESPONSE_TIMEOUT: u32 = 100;
+		const NEW_KEYGEN_RESPONSE_TIMEOUT: u32 = 101;
+		const NEW_KEYGEN_SLASH_AMOUNT: FlipBalance = 102;
+
+		// Check that the default values are different from the new ones
+		assert_ne!(
+			ThresholdSignatureResponseTimeout::<Test, Instance1>::get(),
+			BlockNumberFor::<Test>::from(NEW_THRESHOLD_SIGNATURE_RESPONSE_TIMEOUT)
 		);
-		assert_noop!(
-			EvmThresholdSigner::set_keygen_slash_amount(RuntimeOrigin::signed(ALICE), 1),
-			sp_runtime::traits::BadOrigin,
+		assert_ne!(
+			KeygenResponseTimeout::<Test, Instance1>::get(),
+			BlockNumberFor::<Test>::from(NEW_KEYGEN_RESPONSE_TIMEOUT)
 		);
-		assert_noop!(
-			EvmThresholdSigner::set_threshold_signature_timeout(RuntimeOrigin::signed(ALICE), 1),
-			sp_runtime::traits::BadOrigin,
+		assert_ne!(KeygenSlashAmount::<Test, Instance1>::get(), NEW_KEYGEN_SLASH_AMOUNT);
+
+		// Update all config items
+		let updates = vec![
+			PalletConfigUpdate::ThresholdSignatureResponseTimeout {
+				new_timeout: NEW_THRESHOLD_SIGNATURE_RESPONSE_TIMEOUT,
+			},
+			PalletConfigUpdate::KeygenResponseTimeout { new_timeout: NEW_KEYGEN_RESPONSE_TIMEOUT },
+			PalletConfigUpdate::KeygenSlashAmount { amount_to_slash: NEW_KEYGEN_SLASH_AMOUNT },
+		];
+		for update in updates.clone() {
+			assert_ok!(EvmThresholdSigner::update_pallet_config(
+				OriginTrait::root(),
+				update.clone()
+			));
+			// Check that the events were emitted
+			System::assert_has_event(RuntimeEvent::EvmThresholdSigner(
+				crate::Event::PalletConfigUpdated { update },
+			));
+		}
+
+		// Check that the new values were set
+		assert_eq!(
+			ThresholdSignatureResponseTimeout::<Test, Instance1>::get(),
+			BlockNumberFor::<Test>::from(NEW_THRESHOLD_SIGNATURE_RESPONSE_TIMEOUT)
 		);
+		assert_eq!(
+			KeygenResponseTimeout::<Test, Instance1>::get(),
+			BlockNumberFor::<Test>::from(NEW_KEYGEN_RESPONSE_TIMEOUT)
+		);
+		assert_eq!(KeygenSlashAmount::<Test, Instance1>::get(), NEW_KEYGEN_SLASH_AMOUNT);
+
+		// Make sure that only governance can update the config
+		for update in updates {
+			assert_noop!(
+				EvmThresholdSigner::update_pallet_config(OriginTrait::signed(ALICE), update),
+				sp_runtime::traits::BadOrigin
+			);
+		}
 	});
 }