+++ state-chain/pallets/cf-tokenholder-governance/src/mock.rs
@@ -7,8 +7,6 @@ use cf_traits::{
 use codec::{Decode, Encode};
 use frame_support::{derive_impl, parameter_types, traits::HandleLifetime};
 use frame_system as system;
-use sp_core::H256;
-use sp_runtime::traits::{BlakeTwo256, IdentityLookup};
 
 use system::pallet_prelude::BlockNumberFor;
 
@@ -25,8 +23,6 @@ frame_support::construct_runtime!(
 );
 
 parameter_types! {
-	pub const BlockHashCount: u64 = 250;
-	pub const SS58Prefix: u8 = 42;
 	pub const VotingPeriod: BlockNumberFor<Test> = 10;
 	pub const ProposalFee: Balance = 100;
 	pub const EnactmentDelay: BlockNumberFor<Test> = 20;
@@ -34,29 +30,7 @@ parameter_types! {
 
 #[derive_impl(frame_system::config_preludes::TestDefaultConfig as frame_system::DefaultConfig)]
 impl system::Config for Test {
-	type BaseCallFilter = frame_support::traits::Everything;
-	type BlockWeights = ();
-	type BlockLength = ();
-	type DbWeight = ();
-	type RuntimeOrigin = RuntimeOrigin;
-	type RuntimeCall = RuntimeCall;
-	type Nonce = u64;
-	type Hash = H256;
-	type Hashing = BlakeTwo256;
-	type AccountId = AccountId;
-	type Lookup = IdentityLookup<Self::AccountId>;
 	type Block = Block;
-	type RuntimeEvent = RuntimeEvent;
-	type BlockHashCount = BlockHashCount;
-	type Version = ();
-	type PalletInfo = PalletInfo;
-	type AccountData = ();
-	type OnNewAccount = ();
-	type OnKilledAccount = ();
-	type SystemWeightInfo = ();
-	type SS58Prefix = SS58Prefix;
-	type OnSetCode = ();
-	type MaxConsumers = frame_support::traits::ConstU32<5>;
 }
 
 impl_mock_chainflip!(Test);